Extract various Tools into the .util.cli package
Slim down the overwhelming size of the SolrCLI.java class
