+++ solr/core/src/java/org/apache/solr/util/cli/Tool.java
@@ -1,0 +1,13 @@
+package org.apache.solr.util.cli;
+
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+
+/** Defines the interface to a Solr tool that can be run from the command-line app. */
+public interface Tool {
+    String getName();
+
+    Option[] getOptions();
+
+    int runTool(CommandLine cli) throws Exception;
+}
