+++ solr/core/src/java/org/apache/solr/util/cli/ZkRmTool.java
@@ -1,0 +1,88 @@
+package org.apache.solr.util.cli;
+
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.solr.client.solrj.SolrServerException;
+import org.apache.solr.common.cloud.SolrZkClient;
+import org.apache.solr.util.CLIO;
+import org.apache.solr.util.SolrCLI;
+
+import java.io.PrintStream;
+import java.util.Locale;
+import java.util.concurrent.TimeUnit;
+
+public class ZkRmTool extends ToolBase {
+
+    public ZkRmTool() {
+        this(CLIO.getOutStream());
+    }
+
+    public ZkRmTool(PrintStream stdout) {
+        super(stdout);
+    }
+
+    @Override
+    public Option[] getOptions() {
+        return new Option[]{
+                Option.builder("path")
+                        .argName("path")
+                        .hasArg()
+                        .required(true)
+                        .desc("Path to remove.")
+                        .build(),
+                SolrCLI.OPTION_RECURSE,
+                SolrCLI.OPTION_ZKHOST,
+                SolrCLI.OPTION_VERBOSE
+        };
+    }
+
+    @Override
+    public String getName() {
+        return "rm";
+    }
+
+    @Override
+    public void runImpl(CommandLine cli) throws Exception {
+        SolrCLI.raiseLogLevelUnlessVerbose(cli);
+        String zkHost = SolrCLI.getZkHost(cli);
+
+        if (zkHost == null) {
+            throw new IllegalStateException(
+                    "Solr at "
+                            + cli.getOptionValue("zkHost")
+                            + " is running in standalone server mode, 'zk rm' can only be used when running in SolrCloud mode.\n");
+        }
+        String target = cli.getOptionValue("path");
+        Boolean recurse = Boolean.parseBoolean(cli.getOptionValue("recurse"));
+
+        String znode = target;
+        if (target.toLowerCase(Locale.ROOT).startsWith("zk:")) {
+            znode = target.substring(3);
+        }
+        if (znode.equals("/")) {
+            throw new SolrServerException("You may not remove the root ZK node ('/')!");
+        }
+        echoIfVerbose("\nConnecting to ZooKeeper at " + zkHost + " ...", cli);
+        try (SolrZkClient zkClient =
+                     new SolrZkClient.Builder()
+                             .withUrl(zkHost)
+                             .withTimeout(30000, TimeUnit.MILLISECONDS)
+                             .build()) {
+            if (recurse == false && zkClient.getChildren(znode, null, true).size() != 0) {
+                throw new SolrServerException(
+                        "ZooKeeper node " + znode + " has children and recurse has NOT been specified.");
+            }
+            echo(
+                    "Removing ZooKeeper node "
+                            + znode
+                            + " from ZooKeeper at "
+                            + zkHost
+                            + " recurse: "
+                            + Boolean.toString(recurse));
+            zkClient.clean(znode);
+        } catch (Exception e) {
+            SolrCLI.log.error("Could not complete rm operation for reason: ", e);
+            throw (e);
+        }
+    }
+} // End RmTool class
